+++ Runtime/CreateAssetMenuConstants.cs
@@ -11,6 +11,7 @@ namespace Zor.RandomGenerators
 
 		public const string BatesDistributionFolder = ContinuousDistributionsFolder + "Bates Distribution/";
 		public const string ExponentialDistributionFolder = ContinuousDistributionsFolder + "Exponential Distribution/";
+		public const string ExtremeValueDistributionFolder = ContinuousDistributionsFolder + "Extreme Value Distribution/";
 		public const string GammaDistributionFolder = ContinuousDistributionsFolder + "Gamma Distribution/";
 		public const string IrwinHallDistributionFolder = ContinuousDistributionsFolder + "Irwin-Hall Distribution/";
 		public const string NormalDistributionFolder = ContinuousDistributionsFolder + "Normal Distribution/";