+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Extreme Value distribution.
+
 ## [1.3.1] - 2021-07-23
 
 ### Fixed